разделение на модули main, utils, data
